+++ build_runner/lib/src/generate/build_impl.dart
@@ -92,14 +92,20 @@ Future<BuildResult> build(
   final buildPhases = await createBuildPhases(
       targetGraph, builders, builderConfigOverrides, isReleaseBuild ?? false);
 
-  var result = await singleBuild(environment, options, buildPhases);
+  BuildResult result;
+  if (buildPhases.isEmpty) {
+    _logger.severe('Nothing can be built, yet a build was requested.');
+    result = new BuildResult(BuildStatus.failure, []);
+  } else {
+    result = await _singleBuild(environment, options, buildPhases);
+  }
 
   await terminator.cancel();
   await options.logListener.cancel();
   return result;
 }
 
-Future<BuildResult> singleBuild(BuildEnvironment environment,
+Future<BuildResult> _singleBuild(BuildEnvironment environment,
     BuildOptions options, List<BuildPhase> buildPhases) async {
   var buildDefinition =
       await BuildDefinition.prepareWorkspace(environment, options, buildPhases);