+++ build_runner/lib/src/generate/watch_impl.dart
@@ -92,8 +92,15 @@ Future<ServeHandler> watch(
   final buildPhases = await createBuildPhases(
       targetGraph, builders, builderConfigOverrides, isReleaseBuild ?? false);
 
+  if (buildPhases.isEmpty) {
+    _logger.severe('Nothing can be built, yet a build was requested.');
+    await terminator.cancel();
+    await options.logListener.cancel();
+    return null;
+  }
+
   var watch =
-      runWatch(environment, options, buildPhases, terminator.shouldTerminate);
+      _runWatch(environment, options, buildPhases, terminator.shouldTerminate);
 
   // ignore: unawaited_futures
   watch.buildResults.drain().then((_) async {
@@ -112,7 +119,7 @@ Future<ServeHandler> watch(
 ///
 /// The [BuildState.buildResults] stream will end after the final build has been
 /// run.
-WatchImpl runWatch(BuildEnvironment environment, BuildOptions options,
+WatchImpl _runWatch(BuildEnvironment environment, BuildOptions options,
         List<BuildPhase> buildPhases, Future until) =>
     new WatchImpl(environment, options, buildPhases, until,
         options.rootPackageFilesWhitelist.map((g) => new Glob(g)));